requirements.txt
attrs==21.2.0
bcrypt==3.2.0
certifi==2021.5.30
cffi==1.14.6
charset-normalizer==2.0.5
cryptography==3.4.8
distro==1.6.0
docker==5.0.2
docker-compose==1.29.2
dockerpty==0.4.1
docopt==0.6.2
greenlet==1.1.1
idna==3.2
iniconfig==1.1.1
jsonschema==3.2.0
lxml==4.6.3
multitasking==0.0.9
numpy==1.21.2
packaging==21.0
pandas==1.3.3
paramiko==2.7.2
pluggy==1.0.0
psycopg2==2.9.1
py==1.10.0
pycparser==2.20
PyNaCl==1.4.0
pyparsing==2.4.7
pyrsistent==0.18.0
pytest==6.2.5
python-dateutil==2.8.2
python-decouple==3.4
python-dotenv==0.19.0
pytz==2021.1
PyYAML==5.4.1
requests==2.26.0
six==1.16.0
SQLAlchemy==1.4.23
texttable==1.6.4
toml==0.10.2
urllib3==1.26.6
websocket-client==0.59.0
yfinance==0.1.63
